Get the merge going
