+++ kernel/Makefile
@@ -1,7 +1,7 @@
-# Makefile for the jk1 kernel.
+# Makefile for the OS Zin kernel.
 
 # Note: if `make' fails with "Error 1" after
-# 	[ GRUB ]  jk1.iso
+# 	[ GRUB ]  OSZin.iso
 # Then you can view grub-mkrescue.log which contains the redirected output of
 # `grub-mkrescue'. By default, that file is deleted if `grub-mkrescue' ran
 # without errors.
@@ -13,9 +13,6 @@
 # If you want to know more, read the Make Manual at http://www.gnu.org/:
 # 	https://www.gnu.org/software/make/manual/make.pdf.
 
-# Target architecture.
-ARCH ?= ia32
-ARCHDIR := arch/$(ARCH)
 
 # Cross-compiler prefix.
 CROSS_PREFIX ?= i686-elf-
@@ -27,21 +24,20 @@ SYSROOT := ../sysroot
 CC = @echo " [ CC  ] " $@ ; $(CROSS_PREFIX)gcc
 CXX = @echo " [ CC  ] " $@ ; $(CROSS_PREFIX)g++
 CFLAGS := -c -ffreestanding -Wall -Wextra -Werror
-CPPFLAGS := -Iinclude/ --sysroot=$(SYSROOT) -isystem=/usr/include -D__is_kernel -fpermissive
-
+CPPFLAGS := -Iinclude/ --sysroot=$(SYSROOT) -isystem=/usr/include -D__is_kernel
 # Assembler
 AS = @echo " [ AS  ] " $@ ; $(CROSS_PREFIX)gcc
 ASFLAGS := $(CFLAGS)
 
 # Linker.
 LD = @echo " [ LD  ] " $@ ; $(CROSS_PREFIX)g++
-LDFLAGS := -ffreestanding -Wall -Wextra -Werror -T $(ARCHDIR)/linker.ld\
+LDFLAGS := -ffreestanding -Wall -Wextra -Werror -T linker.ld\
 	--sysroot=$(SYSROOT)
 LDLIBS := -lk -lgcc -nostdlib
 
 # Source files to compile.
-SRC := $(wildcard src/*.[sc] src/*/*.[sc])
-SRC += $(wildcard src/*.cc src/*/*.cc)
+SRC := $(wildcard ./*.[sc] ./*/*.[sc])
+SRC += $(wildcard ./*.cc ./*/*.cc)
 
 # Object files.
 OBJ := $(subst .cc,.o,$(SRC))
@@ -59,15 +55,10 @@ OUTPUT := kernel
 # Where to install the kernel.
 INSTALLDIR := $(SYSROOT)/boot
 
-# Include Makefile for the target arch.
-include $(ARCHDIR)/Makefile.inc
-
-# Phony targets.
-.PHONY: all install clean
 
 all: $(OUTPUT) $(ISOFILE)
 
-$(OUTPUT): $(ARCH_OBJ) $(OBJ)
+$(OUTPUT): $(OBJ)
 	$(LD) $(LDFLAGS) $^ -o $@ $(LDLIBS)
 
 $(ISOFILE): $(OUTPUT) $(ISODIR)/boot/grub | install
@@ -85,4 +76,4 @@ $(INSTALLDIR)/$(OUTPUT): $(OUTPUT)
 	@echo " [ CP  ] " $< ; cp $< $@
 
 clean:
-	rm -rf $(wildcard src/*.o src/*/*.o $(ARCHDIR)/*.o $(ARCHDIR)/*/*.o) $(OUTPUT) $(ISOFILE)
+	rm -rf $(wildcard ./*.o ./*/*.o) $(OUTPUT) $(ISOFILE)
