+++ kernel/grub.cfg
@@ -1,4 +1,3 @@
-
 set default=0
 set timeout=0
 
@@ -6,6 +5,6 @@ set menu_color_normal=dark-gray/light-gray
 set menu_color_highlight=green/light-gray
 set color_normal=dark-gray/black
 
-menuentry "OSZin" {
+menuentry "OS Zin" {
 	multiboot /boot/kernel
 }
