--- kernel/include/arch/ia32/modules/IDT.hpp
@@ -1,78 +1,0 @@
-#ifndef IDT_HPP
-#define IDT_HPP
-#include <stdint.h>
-extern "C" {
-extern void isr0 (void);
-extern void isr1 (void);
-extern void isr2 (void);
-extern void isr3 (void);
-extern void isr4 (void);
-extern void isr5 (void);
-extern void isr6 (void);
-extern void isr7 (void);
-extern void isr8 (void);
-extern void isr9 (void);
-extern void isr10 (void);
-extern void isr11 (void);
-extern void isr12 (void);
-extern void isr13 (void);
-extern void isr14 (void);
-extern void isr15 (void);
-extern void isr16 (void);
-extern void isr17 (void);
-extern void isr18 (void);
-extern void isr19 (void);
-extern void isr20 (void);
-extern void isr21 (void);
-extern void isr22 (void);
-extern void isr23 (void);
-extern void isr24 (void);
-extern void isr25 (void);
-extern void isr26 (void);
-extern void isr27 (void);
-extern void isr28 (void);
-extern void isr29 (void);
-extern void isr30 (void);
-extern void isr31 (void);
-extern void irq0 (void);
-extern void irq1 (void);
-extern void irq2 (void);
-extern void irq3 (void);
-extern void irq4 (void);
-extern void irq5 (void);
-extern void irq6 (void);
-extern void irq7 (void);
-extern void irq8 (void);
-extern void irq9 (void);
-extern void irq10 (void);
-extern void irq11 (void);
-extern void irq12 (void);
-extern void irq13 (void);
-extern void irq14 (void);
-extern void irq15 (void);
-}
-
-typedef struct {
-	uint32_t ds;
-	uint32_t edi, esi, ebp, useless_value, ebx, edx, ecx, eax;
-	uint32_t int_no, err_code;
-	uint32_t eip, cs, eflags, esp, ss;
-} registers_t;
-
-
-typedef struct {
-    uint32_t gs, fs, es, ds;
-    uint32_t edi, esi, ebp, esp, ebx, edx, ecx, eax;
-    uint32_t int_no, err_code;
-    uint32_t eip, cs, eflags, useresp, userss;
-}  interrupt_context;
-
-typedef void (*idt_interruptHandler_t)(interrupt_context *);
-
-extern idt_interruptHandler_t isrHandlers[32];
-extern idt_interruptHandler_t irqHandlers[16];
-
-void idt_initialize();
-void setHandler(int type, int index, idt_interruptHandler_t handler);
-
-#endif
