--- kernel/include/modules/Keyboard.hpp
@@ -1,12 +1,0 @@
-#ifndef KEYBOARD_H
-#define KEYBOARD_H
-
-class Keyboard {
-  public:
-    Keyboard();  //Contructor needed, but can't do anything :/
-    void init();
-  private:
-
-};
-
-#endif
