+++ kernel/main.c
@@ -1,0 +1,4 @@
+#include <multiboot.h>
+
+void kernel_main(multiboot_info *mbt, unsigned int magic) {
+}
