--- sysroot/boot/grub/grub.cfg
@@ -1,11 +1,0 @@
-
-set default=0
-set timeout=0
-
-set menu_color_normal=dark-gray/light-gray
-set menu_color_highlight=green/light-gray
-set color_normal=dark-gray/black
-
-menuentry "OSZin" {
-	multiboot /boot/kernel
-}
